pyproject.toml
[project]
name = "reimagine-movies"
version = "0.1.0"
description = "AI-powered narrative transformation system using multi-agent microservices"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.0.0",
    "pydantic>=2.0.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.25.0",
    "rich>=13.0.0",
    "pyyaml>=6.0.0",
    "tiktoken>=0.5.0",
]
